Fix several builder creation issues
To get a project builder RemoteBuildEngineManager first gets a build
engine and then it gets the project builder from it. The process of
getting the project builder was done inside the engine lock and that
was causing a lot of thread contention that lead to a collapse of
builder creation. To avoid this problem builders are now created
outside of the engine lock. With this change there is now the
possibility of the engine being shutdown before the builder can be
obtained from it. If that happens a recursive call is made to get a
new builder from a new engine.

This patch also improves the handling of engine shutdown. The project
builder creation logic now ensures that builders have a reference
count of 1 when they are created and registerd in the builders list.
This was done before in three steps: create -> register builder -> add
reference. In some cases this lead to the engine to be disposed before
the reference was added.

Fixes VSTS #841060 - Hang when opening a file and lots of background threads